refactor(playground): extract inline extensions from e2e workspace configs
Extract the 100-line inline markdown materializer from opensidian-e2e into
a reusable createOpensidianMaterializer factory. Add createCliUnlock to
@epicenter/cli for the shared CLI session unlock pattern.

- opensidian-e2e config: 184 → 55 lines
- tab-manager-e2e config: 65 → 62 lines (inline unlock → createCliUnlock)
- Standardize opensidian-e2e persistence to filesystemPersistence
- Switch test to use filesystemPersistence for consistency
