+++ packages/cli/src/index.ts
@@ -1,6 +1,7 @@
 /** @module @epicenter/cli — Public API for the Epicenter CLI package. */
 
-export { createCLI, resolveEpicenterHome } from './cli';
 export { createAuthApi, type AuthApi } from './auth/api';
-export { loadConfig, type LoadConfigResult } from './load-config';
 export { createSessionStore, type AuthSession } from './auth/store';
+export { createCLI, resolveEpicenterHome } from './cli';
+export { createCliUnlock } from './extensions';
+export { loadConfig, type LoadConfigResult } from './load-config';
