+++ playground/opensidian-e2e/epicenter.config.test.ts
@@ -17,7 +17,7 @@ import { mkdir, rm } from 'node:fs/promises';
 import { join } from 'node:path';
 import { generateId } from '@epicenter/workspace';
 import { toMarkdown } from '@epicenter/workspace/extensions/materializer/markdown';
-import { persistence } from '@epicenter/workspace/extensions/persistence/sqlite';
+import { filesystemPersistence } from '@epicenter/workspace/extensions/persistence/sqlite';
 import { createWorkspace } from '@epicenter/workspace';
 import { opensidianDefinition } from 'opensidian/workspace';
 import { pushFromMarkdown } from './push-from-markdown';
@@ -33,8 +33,9 @@ function dbPath(id: string) {
 
 /** Create a workspace client with filesystem persistence for testing. */
 function createTestClient() {
-	return createWorkspace(opensidianDefinition).withWorkspaceExtension('persistence', (ctx) =>
-		persistence(ctx, { filePath: dbPath(opensidianDefinition.id) }),
+	return createWorkspace(opensidianDefinition).withExtension(
+		'persistence',
+		filesystemPersistence({ filePath: dbPath(opensidianDefinition.id) }),
 	);
 }
 
@@ -142,8 +143,11 @@ describe('e2e: opensidian pushFromMarkdown', () => {
 	const IMPORT_FILES_DIR = join(IMPORT_DIR, 'files');
 
 	function createImportClient() {
-	return createWorkspace(opensidianDefinition).withWorkspaceExtension('persistence', (ctx) =>
-			persistence(ctx, { filePath: join(IMPORT_PERSISTENCE, 'opensidian.db') }),
+		return createWorkspace(opensidianDefinition).withExtension(
+			'persistence',
+			filesystemPersistence({
+				filePath: join(IMPORT_PERSISTENCE, 'opensidian.db'),
+			}),
 		);
 	}
